requirements.txt
pytest==3.4.0
scipy==1.1.0
pillow==5.2.0
pandas==0.22.0; python_version < '3.7'
pandas==1.3.5; python_version >= '3.7'
numpy==1.19.5; python_version < '3.7'
numpy==1.20.1; python_version >= '3.7'
regex==2018.2.3
typing==3.7.4.3
